Consolidate code architecture, namely animatedsprite and animations into sprites module.
  * Add sprites module
  * Rename util module to resources
  * Move animatedsprite module into sprites module
  * Move animations module into sprites module
  * Add filehandlers module for all the filehandlers Resource uses. Move file handling functions from util/resources to filehandlers module.
  * Rename config_fromfp to load_ini
  * Resource now has a FROM_EXTENSION constant (dict) which allows you to exchange a file extension for a filehandler functoin
  * Correct code, doctests and tests to succeed with all of the changes aformentioned
