+++ hypatia/constants.py
@@ -15,14 +15,14 @@ direction and action. The benefit of this model is:
     velocity to direction
 
 Example:
-    >>> from hypatia import animations
-    >>> sprite = animations.Walkabout('debug')
+    >>> from hypatia import sprites
+    >>> sprite = sprites.Walkabout('debug')
     >>> sprite.animations[Action.walk][Direction.east]
     <AnimatedSprite sprite(in ... groups)>
 
 See Also:
    *  :attribute:`actor.Actor.direction`
-   *  :attribute:`animations.Walkabout.direction`
+   *  :attribute:`sprites.Walkabout.direction`
 
 """
 
@@ -312,7 +312,7 @@ class Action(enum.Enum):
             has velocity.
 
     See Also:
-        :class:`animations.Walkabout`
+        :class:`sprites.Walkabout`
 
     """
