+++ hypatia/render.py
@@ -19,7 +19,6 @@ import pygame
 import pyganim
 from pygame.locals import *
 
-from hypatia import util
 from hypatia import constants
