+++ hypatia/resources.py
@@ -1,27 +1,21 @@
 # This module is part of Hypatia and is released under the
 # MIT license: http://opensource.org/licenses/MIT
 
-"""These are utilities which are commonly utilized
+"""Manage assets and resources.
+
+These are utilities which are commonly utilized
 by all modules in Hypatia. It serves for the ugly,
 underlying components of miscellaneous actions which
 assist other modules, and does not do much on its own.
+
+need to put file handlers in its own module
 
 """
 
 import os
 import zipfile
-from io import BytesIO
 
-try:
-    import ConfigParser as configparser
-    from cStringIO import StringIO
-
-except ImportError:
-    import configparser
-    from io import StringIO
-
-import pygame
-from hypatia.animatedsprite import AnimatedSprite
+from hypatia import filehandlers
 
 
 class Resource(object):
@@ -35,17 +29,25 @@ class Resource(object):
             BytesIO, PygAnim, or ConfigParser objects.
 
     Example:
-        >>> from hypatia import animatedsprite as anim
+        >>> from hypatia import sprites
         >>> resource = Resource('walkabouts', 'debug')
         >>> 'walk_north.gif' in resource
         True
-        >>> isinstance(resource['walk_north.gif'], anim.AnimatedSprite)
+        >>> isinstance(resource['walk_north.gif'],
+        ...            sprites.AnimatedSprite)
         True
         >>> resource = Resource('scenes', 'debug')
         >>> resource['tilemap.txt'].startswith('debug')
         True
 
     """
+
+    FROM_EXTENSION = {
+                     '.ini': filehandlers.load_ini,
+                     '.gif': filehandlers.load_gif,
+                     '.png': filehandlers.load_png,
+                     '.txt': filehandlers.load_txt,
+                    }
 
     def __init__(self, resource_category, resource_name):
         """Load a resource ZIP using a category and zip name.
@@ -69,16 +71,9 @@ class Resource(object):
                             resource_category,
                             resource_name
                            )
-
-        # Once files have been collected from the aforementioned
+        # ... Once files have been collected from the aforementioned
         # path, the files will be passed through their respective
         # file_handler, if available for the given file extension.
-        file_handlers = {
-                         '.ini': configparser_fromfp,
-                         '.gif': load_gif,
-                         '.png': load_png,
-                         '.txt': load_txt,
-                        }
 
         # 1. Create a dictionary, where the key is the file name
         # (including extension) and the value is the result
@@ -123,8 +118,9 @@ class Resource(object):
             # if there is a known "handler" for this extension,
             # we want the file data for this file to be the output
             # of said handler
-            if file_extension in file_handlers:
-                file_data = file_handlers[file_extension](files, file_name)
+            if file_extension in Resource.FROM_EXTENSION:
+                file_handler = Resource.FROM_EXTENSION[file_extension]
+                file_data = file_handler(files, file_name)
 
             files[file_name] = file_data
 
@@ -164,128 +160,3 @@ class Resource(object):
                 matching_files[file_name] = file_content
 
         return matching_files or None
-
-
-def load_png(files, file_name):
-    """Return an BytesIO object based on supplied file. This is
-    a file handler for Resource.
-
-    Args:
-        files (dict): Resources files, whereas key is the file name,
-            and the value is the untouched file contents itself.
-        file_name (str): File from "files" to use for making an
-            AnimatedSprite object.
-
-    Returns:
-        AnimatedSprite: --
-
-    See Also:
-        * Resources.__init__()
-        * animations.AnimatedSprite
-
-    """
-
-    return BytesIO(files[file_name])
-
-
-def load_txt(files, file_name):
-    """Return a decoded string based on supplied file. This is
-    a file handler for Resource.
-
-    Args:
-        files (dict): Resource files, whereas key is the file
-            name and the value is the untouched file contents
-            itself.
-        file_name (StR): File from "files" to use for making
-            an animatedSprite object.
-
-    Returns:
-        AnimatedSprite: --
-
-    See Also:
-        * Resources.__init__()
-        * animations.AnimatedSprite
-
-    """
-
-    return files[file_name].decode('utf-8')
-
-
-def load_gif(files, file_name):
-    """Return an AnimatedSprite object based on a bytesio
-    object. This is a file handler.
-
-    Args:
-        files (dict): Resources files, whereas key is the file name,
-            and the value is the untouched file contents itself.
-        file_name (str): File from "files" to use for making an
-            AnimatedSprite object.
-
-    Returns:
-        AnimatedSprite: --
-
-    See Also:
-        * Resources.__init__()
-        * animations.AnimatedSprite
-
-    """
-
-    file_data = files[file_name]
-
-    # NOTE: i used to handle this just in
-    # Resources.__init__()
-    gif_bytesio = BytesIO(file_data)
-
-    # get the corersponding INI which configures our anchor points
-    # for this gif, from the files
-    gif_name_no_ext = os.path.splitext(file_name)[0]
-
-    try:
-        anchor_ini_name = gif_name_no_ext + '.ini'
-        anchor_config_ini = files[anchor_ini_name]
-
-        # if the INI file has not already been parsed into
-        # ConfigParser object, we'll do that now, so we
-        # can accurately construct our AnimatedSprite.
-        try:
-            anchor_config_ini.sections()
-        except AttributeError:
-            anchor_config_ini = configparser_fromfp(files, anchor_ini_name)
-
-    except KeyError:
-        anchor_config_ini = None
-
-    return AnimatedSprite.from_file(gif_bytesio, anchor_config_ini)
-
-
-def configparser_fromfp(files, file_name):
-    """Return a ConfigParser object based on a bytesio
-    object. This is a file handler.
-
-    Args:
-        files (dict): Resources files, whereas key is the file name,
-            and the value is a BytesIO object of said file.
-        file_name (str): File from "files" to use for making a
-            ConfigParser object.
-
-    Returns:
-        ConfigParser: --
-
-    See Also:
-        Resources.__init__()
-
-    """
-
-    file_data = files[file_name]
-
-    # i used to do this in Resources.__init__()
-    file_data = file_data.decode('utf-8')
-
-    file_data = StringIO(file_data)
-    config = configparser.ConfigParser()
-
-    # NOTE: this still works in python 3, though it was
-    # replaced by config.read_file()
-    config.readfp(file_data)
-
-    return config
